+++ test/misc_rules/BUILD
@@ -9,7 +9,6 @@ filegroup(
     provides = {
         "txt": ":txt",
     },
-    deps = [":not_require_filegroup"],
 )
 
 filegroup(
